+++ projectroles/views.py
@@ -89,7 +89,6 @@ APP_SETTING_SCOPE_PROJECT = SODAR_CONSTANTS['APP_SETTING_SCOPE_PROJECT']
 
 # Local constants
 APP_NAME = 'projectroles'
-SEARCH_REGEX = re.compile(r'^[a-zA-Z0-9.:\-_\s\t]+$')
 ALLOWED_CATEGORY_URLS = ['detail', 'create', 'update', 'ajax_star']
 KIOSK_MODE = getattr(settings, 'PROJECTROLES_KIOSK_MODE', False)
 
@@ -549,12 +548,6 @@ class ProjectSearchView(LoginRequiredMixin, TemplateView):
             return redirect('home')
 
         context = self.get_context_data(*args, **kwargs)
-
-        # Check input, redirect if unwanted characters are found
-        if not bool(re.match(SEARCH_REGEX, context['search_input'])):
-            messages.error(request, 'Please check your search input')
-            return redirect('home')
-
         return super().render_to_response(context)